Ajout des controlleurs media et member et modification des model

+++ src/main/java/org/dta/alexendrie/core/JpaRepositoryImpl.java
@@ -36,7 +36,7 @@ public abstract class JpaRepositoryImpl<T extends Model> implements JpaRepositor
     }
 
 	@Transactional
-    public T findOne(int id) {
+    public T findOne(long id) {
         return em.find(entityClass, id);
     }
